feat: add working implementation of react hooks
After hours of debugging, I got the react hooks working now by using
custom events. A major pain in the ass due to Discord.py heavily lacking
on documentation for registering events.

Next step is to implement commands for adding and removing hooks

NOTE TO FUTURE ME:
Discord.py implicitly adds "on_" to the start of whatever event you try
to dispatch. This isn't anywhere in the docs so I had to literally look
at the source to find this.