Add ChatHook framework (#16)
Closes #13 

* fix: add -u flag to python to rm buffering on stdout & stderr

* feat: add new inspection dev commands + switch to per-cog error handling

* feat: add working implementation of react hooks

After hours of debugging, I got the react hooks working now by using
custom events. A major pain in the ass due to Discord.py heavily lacking
on documentation for registering events.

Next step is to implement commands for adding and removing hooks

NOTE TO FUTURE ME:
Discord.py implicitly adds "on_" to the start of whatever event you try
to dispatch. This isn't anywhere in the docs so I had to literally look
at the source to find this.

* style(QuadChat): rm debug print statements

* fix(QuadDev): fix reload_all command not functioning

I had forgotten to adjust the path to the cogs after rearranging the
project structure sometime in the past. I should consider a more dynamic
mechanism to set that path at startup.

Ref: #13

* feat(QuadReact): make hooks polymorphic

I created a standard interface for the Hooks and used dynamic dispatch
to eliminate visible runtime typechecking for the response type.
I hope to add arbitrary callbacks in the future as well.

* style(QuadDev): change reload log to say reloading extension

* ref(QuadReact): swap implementation to use multiple regex patterns

* style(QuadChat): rename default hooks constant

* feat(QuadReact): add EventHook scaffolding for triggering arbitrary func

* feat(QuadReact): add EventHooks for ++/-- elo system

* style(QuadReact): add comments + formatting

* feat(QuadReact): prevent users from targeting bots with chathooks

* feat(QuadReact): add show_hooks command for displaying active ChatHooks